Improve integrations tests to have more platformss
